Fix missing "Use Skija" button
The button to activate Skija rendering is not visible in some cases.
Reason for this is
- 1. The button is resized to height 0 if not sufficient space
- 2. Windows in maximized mode is rendered not considering the task bar
and therefore part of the window is not visible.

Fix for 1. Set grid layout data to not grab excess space
Fix for 2. As a workaround: Move Button to top. As a result, a scroll
bar is not displayed but this scroll bar is not needed anyway.

Also but code for tabFolder creation/button creation and mac workaround
in methods to de-clutter constructor.